requirements.txt
scrapy
requests
pycrypto
maraidb
